rename to favourites.js
